fixed issue with assign node when left side has a filtering expression

+++ src/Value.cpp
@@ -4,10 +4,10 @@ using namespace dplsrc;
 
 std::string Value::toString() const {
     if (is<INT>()) {
-        return std::to_string(get<long>());
+        return std::to_string(get<INT>());
     }
     if (is<FLOAT>()) {
-        std::string doubleStr = std::to_string(get<double>());
+        std::string doubleStr = std::to_string(get<FLOAT>());
 
         // Remove trailing zeros
         doubleStr.erase(doubleStr.find_last_not_of('0') + 1, std::string::npos);
@@ -185,11 +185,11 @@ bool Value::operator==(const Value& other) const {
         (other.is<INT>() || other.is<FLOAT>() || other.is<BOOL>())) {
         Value::FLOAT val1 = is<INT>()
                                 ? get<INT>()
-                                : (is<FLOAT>() ? get<FLOAT>() : static_cast<double>(get<BOOL>()));
+                                : (is<FLOAT>() ? get<FLOAT>() : static_cast<FLOAT>(get<BOOL>()));
         Value::FLOAT val2 =
             other.is<INT>()
                 ? other.get<INT>()
-                : (other.is<FLOAT>() ? other.get<FLOAT>() : static_cast<double>(other.get<BOOL>()));
+                : (other.is<FLOAT>() ? other.get<FLOAT>() : static_cast<FLOAT>(other.get<BOOL>()));
         return val1 == val2;
     }
     if (is<STR>() && other.is<STR>()) {
@@ -254,11 +254,11 @@ bool Value::operator<(const Value& other) const {
         (other.is<INT>() || other.is<FLOAT>() || other.is<BOOL>())) {
         Value::FLOAT val1 = is<INT>()
                                 ? get<INT>()
-                                : (is<FLOAT>() ? get<FLOAT>() : static_cast<double>(get<BOOL>()));
+                                : (is<FLOAT>() ? get<FLOAT>() : static_cast<FLOAT>(get<BOOL>()));
         Value::FLOAT val2 =
             other.is<INT>()
                 ? other.get<INT>()
-                : (other.is<FLOAT>() ? other.get<FLOAT>() : static_cast<double>(other.get<BOOL>()));
+                : (other.is<FLOAT>() ? other.get<FLOAT>() : static_cast<FLOAT>(other.get<BOOL>()));
         return val1 < val2;
     }
     if (is<STR>() && other.is<STR>()) {
@@ -335,8 +335,8 @@ Value Value::operator+(const Value& other) const {
     const Value& val2 = other;
 
     if ((val1.is<INT>() || val1.is<BOOL>()) && (val2.is<INT>() || val2.is<BOOL>())) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value(int1 + int2);
     }
@@ -344,11 +344,11 @@ Value Value::operator+(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(float1 + float2);
     }
@@ -384,8 +384,8 @@ Value Value::operator-(const Value& other) const {
     const Value& val2 = other;
 
     if ((val1.is<INT>() || val1.is<BOOL>()) && (val2.is<INT>() || val2.is<BOOL>())) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value(int1 - int2);
     }
@@ -393,11 +393,11 @@ Value Value::operator-(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(float1 - float2);
     }
@@ -418,8 +418,8 @@ Value Value::operator*(const Value& other) const {
     bool isVal2Int = val2.is<INT>() || val2.is<BOOL>();
 
     if (isVal1Int && isVal2Int) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value(int1 * int2);
     }
@@ -427,11 +427,11 @@ Value Value::operator*(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(float1 * float2);
     }
@@ -440,10 +440,10 @@ Value Value::operator*(const Value& other) const {
         std::string strVal;
 
         if (val1.is<INT>() || val1.is<BOOL>()) {
-            intVal = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
+            intVal = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
             strVal = val2.get<STR>();
         } else {
-            intVal = (val2.get<INT>()) != 0 ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+            intVal = (val2.get<INT>()) != 0 ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
             strVal = val1.get<STR>();
         }
 
@@ -460,10 +460,10 @@ Value Value::operator*(const Value& other) const {
         Value::LIST listVal;
 
         if (val1.is<INT>() || val1.is<BOOL>()) {
-            intVal = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
+            intVal = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
             listVal = val2.get<LIST>();
         } else {
-            intVal = (val2.get<INT>()) != 0 ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+            intVal = (val2.get<INT>()) != 0 ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
             listVal = val1.get<LIST>();
         }
 
@@ -494,8 +494,8 @@ Value Value::operator/(const Value& other) const {
         throw RuntimeException("Cannot divide values with divisor 0");
     }
     if ((val1.is<INT>() || val1.is<BOOL>()) && (val2.is<INT>() || val2.is<BOOL>())) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value::INT(int1 / int2);
     }
@@ -503,11 +503,11 @@ Value Value::operator/(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(float1 / float2);
     }
@@ -530,8 +530,8 @@ Value Value::operator%(const Value& other) const {
         throw RuntimeException("Cannot take remainder of values with divisor 0");
     }
     if ((val1.is<INT>() || val1.is<BOOL>()) && (val2.is<INT>() || val2.is<BOOL>())) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value::INT(int1 % int2);
     }
@@ -539,11 +539,11 @@ Value Value::operator%(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(std::fmod(float1, float2));
     }
@@ -561,8 +561,8 @@ Value Value::pow(const Value& other) const {
     const Value& val2 = other;
 
     if ((val1.is<INT>() || val1.is<BOOL>()) && (val2.is<INT>() || val2.is<BOOL>())) {
-        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<long>(val1.get<BOOL>());
-        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<long>(val2.get<BOOL>());
+        Value::INT int1 = (val1.is<INT>()) ? val1.get<INT>() : static_cast<INT>(val1.get<BOOL>());
+        Value::INT int2 = (val2.is<INT>()) ? val2.get<INT>() : static_cast<INT>(val2.get<BOOL>());
 
         return Value::INT(std::pow(int1, int2));
     }
@@ -570,11 +570,11 @@ Value Value::pow(const Value& other) const {
         Value::FLOAT float1 =
             (val1.is<INT>())
                 ? val1.get<INT>()
-                : ((val1.is<BOOL>()) ? static_cast<double>(val1.get<BOOL>()) : val1.get<FLOAT>());
+                : ((val1.is<BOOL>()) ? static_cast<FLOAT>(val1.get<BOOL>()) : val1.get<FLOAT>());
         Value::FLOAT float2 =
             (val2.is<INT>())
                 ? val2.get<INT>()
-                : ((val2.is<BOOL>()) ? static_cast<double>(val2.get<BOOL>()) : val2.get<FLOAT>());
+                : ((val2.is<BOOL>()) ? static_cast<FLOAT>(val2.get<BOOL>()) : val2.get<FLOAT>());
 
         return Value(std::pow(float1, float2));
     }
@@ -670,10 +670,10 @@ Value Value::colOperations(const Value& val1, const Value& val2,
 
 double Value::getNumericValue() const {
     if (is<BOOL>()) {
-        return static_cast<double>(get<BOOL>());
+        return static_cast<FLOAT>(get<BOOL>());
     }
     if (is<INT>()) {
-        return static_cast<double>(get<INT>());
+        return static_cast<FLOAT>(get<INT>());
     }
     if (is<FLOAT>()) {
         return get<FLOAT>();
